Extend unit tests
